refactor(plotables): extract SciQLopMultiGraphBase from SciQLopLineGraph
Buffer dispatch, dataGuard, and per-column component creation move to the
base class; SciQLopLineGraph becomes a thin derived class that only
supplies the concrete QCPMultiGraph factory. Public API unchanged;
regression suite test_line_graph_regression.py added to snapshot behavior.
